[refactor] 리엑트 쿼리 분리 작업 진행
